refactor: extract renderers and style configs
